add functionality for one-time lock tracking
Each time the rods A or B are locked, they update a boolean variable
that becomes true upon reaching the lock state (extended fully). That
state remains true until the user unlocks the bike rack.

The variable is used mainly for force triggering the alarm when the rod
is being forced to unlock by an external force instead of being unlocked
through the app.